moved data and keras_mnist_3.py to appropriate locations
